Subscriptions will only update if there have been changes to the current subscription map.
When a new subscription is added to the map, it sets a flag that there's an update. The code that does the forwarding (uncommented now) will only be called if that flag is set. Flag is reset after the forwarding is executed.
Fixed bug in map where it would search map by ComponentTypeID but put into map with ComponentTypeID.getComponentID() as the key.